+++ src/components/join/index.js
@@ -2,3 +2,4 @@ export { default as CancelJoinPopup } from "./CancelJoinPopup";
 export { default as JoinHeader } from "./JoinHeader";
 export { default as JoinButton } from "./JoinButton";
 export { default as JoinForm } from "./JoinForm";
+export { default as JoinTextField } from "./JoinTextField";
